requirements.txt
requests==2.28.0
websocket-client==1.3.3
python-telegram-bot==13.12
